+++ src/career_quest/career_quest_triggers.h
@@ -1,118 +1,129 @@
 #pragma once
 
 #include <Adafruit_NeoPixel.h>
+
+#include "actuators/buttons.h"
+#include "actuators/display_screen.h"
+#include "actuators/led/led_animations.h"
+#include "actuators/led/rgb_led.h"
+#include "actuators/speaker.h"
 #include "networking/protocol.h"
 #include "utils/config.h"
 #include "utils/structs.h"
-#include "actuators/display_screen.h"
-#include "actuators/led/rgb_led.h"
-#include "actuators/led/led_animations.h"
-#include "actuators/buttons.h"
-#include "actuators/speaker.h"
 
 class CareerQuestTriggers {
-    public:
-        CareerQuestTriggers(Adafruit_NeoPixel& strip);
+  public:
+    CareerQuestTriggers(Adafruit_NeoPixel& strip);
 
-        void start_s2_p1_sequence();
-        void stop_s2_p1_sequence();
-        void start_s2_p4_light_show();
-        void stop_s2_p4_light_show();
-        void start_s3_p3_display_demo();
-        void stop_s3_p3_display_demo();
-        void start_s7_p4_button_demo();
-        void stop_s7_p4_button_demo();
-        void start_s5_p4_led_visualization();
-        void stop_s5_p4_led_visualization();
-        void stop_all_career_quest_triggers(bool shouldTurnLedsOff);
-        void update();
-        
-        bool isS2P1Active() const { return s2p1Active; }
-        bool isS2P4Active() const { return s2p4Active; }
-        bool isS3P3Active() const { return s3p3Active; }
-        bool isS7P4Active() const { return s7p4Active; }
-        bool isS5P4Active() const { return s5p4Active; }
-        
-        void render_s3_p3_animation();
-        
-        // LED sequence order (back_right → middle_right → etc.) including headlights
-        static constexpr uint8_t s2p1LedSequence[8] = {7, 0, 1, 2, 3, 4, 5, 6}; // Include headlights (2,3)
-        
-        // Color progression (Red → Orange → Yellow → Green → Blue → Purple → Cyan → Magenta)
-        static constexpr uint8_t s2p1ColorSequence[8][3] = {
-            {255, 0, 0},     // Red
-            {255, 127, 0},   // Orange
-            {255, 255, 0},   // Yellow
-            {127, 255, 0},   // Chartreuse
-            {0, 255, 0},     // Green
-            {0, 255, 255},   // Cyan
-            {0, 0, 255},     // Blue
-            {128, 0, 255}    // Purple
-        };
-        
-        // Timing constants
-        static constexpr unsigned long S2P1_UPDATE_INTERVAL = 10; // 10ms updates for faster, smoother fading
-        static constexpr uint8_t S2P1_BRIGHTNESS_STEP = 10; // 10ms updates for faster, smoother fading
-        static constexpr unsigned long S2P4_DURATION = 15000; // 15 seconds total
-        static constexpr unsigned long S2P4_SNAKE_DURATION = 6000; // 6 seconds
-        static constexpr unsigned long S2P4_FLASH_DURATION = 4000; // 4 seconds  
-        static constexpr unsigned long S2P4_BREATHING_DURATION = 5000; // 5 seconds
-        static constexpr unsigned long S2P4_SNAKE_INTERVAL = 200; // 200ms per LED
-        static constexpr unsigned long S2P4_FLASH_INTERVAL = 400; // 400ms per flash
-        static constexpr unsigned long S3P3_UPDATE_INTERVAL = 50; // 50ms updates for smooth scrolling
-        
-    private:
-        Adafruit_NeoPixel& strip;
-        
-        // S2_P1 sequence state
-        bool s2p1Active = false;
-        uint8_t currentLedIndex = 0;
-        uint8_t currentColorIndex = 0;
-        unsigned long lastS2P1Update = 0;
-        
-        // Fade state
-        bool isFadingOut = false;
-        bool isExitFading = false;
-        uint8_t currentBrightness = 0;
-        uint8_t targetBrightness = 255;
-        
-        // S2_P4 light show state
-        bool s2p4Active = false;
-        unsigned long s2p4StartTime = 0;
-        unsigned long lastS2P4Update = 0;
-        uint8_t s2p4Step = 0;
-        bool s2p4ExitFading = false;
-        uint8_t s2p4CurrentBrightness = 255;
-        
-        // S2P4 3-phase light show
-        uint8_t s2p4Phase = 0;  // 0=snake, 1=flash, 2=breathing
-        uint8_t s2p4SnakePosition = 0;
-        uint8_t s2p4FlashColorIndex = 0;
-        bool s2p4SnakeDirection = true; // true=forward, false=reverse
-        unsigned long s2p4PhaseStartTime = 0;
-        
-        // S3_P3 display demo state
-        bool s3p3Active = false;
-        unsigned long lastS3P3Update = 0;
-        unsigned long s3p3StartTime = 0;
-        uint8_t s3p3AnimationStep = 0;
-        
-        // S7_P4 button demo state
-        bool s7p4Active = false;
-        bool s7p4ExitFading = false;
-        uint8_t s7p4CurrentBrightness = 255;
-        unsigned long lastS7P4Update = 0;
-        
-        // S5_P4 IMU LED visualization state
-        bool s5p4Active = false;
-        bool s5p4ExitFading = false;
-        unsigned long lastS5P4Update = 0;
-        static constexpr unsigned long S5P4_UPDATE_INTERVAL = 20; // 20ms for responsive visualization
-        
-        void update_s2_p1_sequence();
-        void update_s2_p4_light_show();
-        void update_s7_p4_button_demo();
-        void update_s5_p4_led_visualization();
+    void start_s2_p1_sequence();
+    void stop_s2_p1_sequence();
+    void start_s2_p4_light_show();
+    void stop_s2_p4_light_show();
+    void start_s3_p3_display_demo();
+    void stop_s3_p3_display_demo();
+    void start_s7_p4_button_demo();
+    void stop_s7_p4_button_demo();
+    void start_s5_p4_led_visualization();
+    void stop_s5_p4_led_visualization();
+    void stop_all_career_quest_triggers(bool shouldTurnLedsOff);
+    void update();
+
+    bool isS2P1Active() const {
+        return s2p1Active;
+    }
+    bool isS2P4Active() const {
+        return s2p4Active;
+    }
+    bool isS3P3Active() const {
+        return s3p3Active;
+    }
+    bool isS7P4Active() const {
+        return s7p4Active;
+    }
+    bool isS5P4Active() const {
+        return s5p4Active;
+    }
+
+    void render_s3_p3_animation();
+
+    // LED sequence order (back_right → middle_right → etc.) including headlights
+    static constexpr uint8_t s2p1LedSequence[8] = {7, 0, 1, 2, 3, 4, 5, 6}; // Include headlights (2,3)
+
+    // Color progression (Red → Orange → Yellow → Green → Blue → Purple → Cyan → Magenta)
+    static constexpr uint8_t s2p1ColorSequence[8][3] = {
+        {255, 0, 0},   // Red
+        {255, 127, 0}, // Orange
+        {255, 255, 0}, // Yellow
+        {127, 255, 0}, // Chartreuse
+        {0, 255, 0},   // Green
+        {0, 255, 255}, // Cyan
+        {0, 0, 255},   // Blue
+        {128, 0, 255}  // Purple
+    };
+
+    // Timing constants
+    static constexpr uint32_t S2P1_UPDATE_INTERVAL = 10;      // 10ms updates for faster, smoother fading
+    static constexpr uint8_t S2P1_BRIGHTNESS_STEP = 10;       // 10ms updates for faster, smoother fading
+    static constexpr uint32_t S2P4_DURATION = 15000;          // 15 seconds total
+    static constexpr uint32_t S2P4_SNAKE_DURATION = 6000;     // 6 seconds
+    static constexpr uint32_t S2P4_FLASH_DURATION = 4000;     // 4 seconds
+    static constexpr uint32_t S2P4_BREATHING_DURATION = 5000; // 5 seconds
+    static constexpr uint32_t S2P4_SNAKE_INTERVAL = 200;      // 200ms per LED
+    static constexpr uint32_t S2P4_FLASH_INTERVAL = 400;      // 400ms per flash
+    static constexpr uint32_t S3P3_UPDATE_INTERVAL = 50;      // 50ms updates for smooth scrolling
+
+  private:
+    Adafruit_NeoPixel& strip;
+
+    // S2_P1 sequence state
+    bool s2p1Active = false;
+    uint8_t currentLedIndex = 0;
+    uint8_t currentColorIndex = 0;
+    uint32_t lastS2P1Update = 0;
+
+    // Fade state
+    bool isFadingOut = false;
+    bool isExitFading = false;
+    uint8_t currentBrightness = 0;
+    uint8_t targetBrightness = 255;
+
+    // S2_P4 light show state
+    bool s2p4Active = false;
+    uint32_t s2p4StartTime = 0;
+    uint32_t lastS2P4Update = 0;
+    uint8_t s2p4Step = 0;
+    bool s2p4ExitFading = false;
+    uint8_t s2p4CurrentBrightness = 255;
+
+    // S2P4 3-phase light show
+    uint8_t s2p4Phase = 0; // 0=snake, 1=flash, 2=breathing
+    uint8_t s2p4SnakePosition = 0;
+    uint8_t s2p4FlashColorIndex = 0;
+    bool s2p4SnakeDirection = true; // true=forward, false=reverse
+    uint32_t s2p4PhaseStartTime = 0;
+
+    // S3_P3 display demo state
+    bool s3p3Active = false;
+    uint32_t lastS3P3Update = 0;
+    uint32_t s3p3StartTime = 0;
+    uint8_t s3p3AnimationStep = 0;
+
+    // S7_P4 button demo state
+    bool s7p4Active = false;
+    bool s7p4ExitFading = false;
+    uint8_t s7p4CurrentBrightness = 255;
+    uint32_t lastS7P4Update = 0;
+
+    // S5_P4 IMU LED visualization state
+    bool s5p4Active = false;
+    bool s5p4ExitFading = false;
+    uint32_t lastS5P4Update = 0;
+    static constexpr uint32_t S5P4_UPDATE_INTERVAL = 20; // 20ms for responsive visualization
+
+    void update_s2_p1_sequence();
+    void update_s2_p4_light_show();
+    void update_s7_p4_button_demo();
+    void update_s5_p4_led_visualization();
 };
 
 extern CareerQuestTriggers careerQuestTriggers;
