+++ src/demos/obstacle_avoider.cpp
@@ -22,7 +22,7 @@ void ObstacleAvoider::update() {
         return;
     }
 
-    unsigned long current_time = millis();
+    uint32_t current_time = millis();
     if (current_time - _lastUpdateTime < _UPDATE_INTERVAL) {
         return; // Maintain update rate
     }
